# cas-spec/target/generated-sources/com/neo/v1/cas/api/NeoServiceV1Api.java
package com.neo.v1.cas.api;

import java.util.List;

import com.neo.v1.cas.model.*;
import com.neo.v1.cas.type.*;

import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;

/**
* Auto generated by ila/neo.
*/
@Validated
// @RestController
public interface NeoServiceV1Api {

	// @GetMapping("/cas/configurations")
	@Validated
	ResponseEntity<CasResponse> getCasConfigurations(@RequestParam(required = false) String segment, @RequestHeader(required = false) String filterByParameter, @RequestHeader(required = false) String unitName);


}