# cas-spec/target/generated-sources/com/neo/v1/cas/model/CasConfigurationData.java
package com.neo.v1.cas.model;

import com.neo.openapi.utils.serialization.*;
import com.neo.openapi.utils.annotation.*;
import com.neo.v1.cas.type.*;
import javax.validation.constraints.*;
import com.fasterxml.jackson.databind.annotation.*;

import java.util.List;
import lombok.AllArgsConstructor;
import lombok.Builder;

/**
* Auto generated by ila/neo 1.2.0
*/
@Builder
@AllArgsConstructor
public class CasConfigurationData  {

	@NotNull
	private List<Configuration> configurations;


  public CasConfigurationData() {
    super();
  }

  /* Getters */

	public List<Configuration> getConfigurations() { return this.configurations; }


  /* Setters */

	public void setConfigurations(List<Configuration> configurations) { this.configurations = configurations; };


  public static class Builder {

    private CasConfigurationData item = new CasConfigurationData();

		public Builder withConfigurations(List<Configuration> configurations) { item.setConfigurations(configurations); return this; };


    public CasConfigurationData build() {
      return item;
    }

  }


}