# cas-spec/target/generated-sources/com/neo/v1/cas/model/CasResponse.java
package com.neo.v1.cas.model;

import com.neo.openapi.utils.serialization.*;
import com.neo.openapi.utils.annotation.*;
import com.neo.v1.cas.type.*;
import javax.validation.constraints.*;
import com.fasterxml.jackson.databind.annotation.*;

import lombok.AllArgsConstructor;
import lombok.Builder;

/**
* Auto generated by ila/neo 1.2.0
*/
@Builder
@AllArgsConstructor
public class CasResponse  {

	@NotNull
	private Meta meta;

	private CasConfigurationData data;


  public CasResponse() {
    super();
  }

  /* Getters */

	public Meta getMeta() { return this.meta; }

	public CasConfigurationData getData() { return this.data; }


  /* Setters */

	public void setMeta(Meta meta) { this.meta = meta; };

	public void setData(CasConfigurationData data) { this.data = data; };


  public static class Builder {

    private CasResponse item = new CasResponse();

		public Builder withMeta(Meta meta) { item.setMeta(meta); return this; };

		public Builder withData(CasConfigurationData data) { item.setData(data); return this; };


    public CasResponse build() {
      return item;
    }

  }


}