# cas-spec/target/generated-sources/com/neo/v1/cas/model/Configuration.java
package com.neo.v1.cas.model;

import com.neo.openapi.utils.serialization.*;
import com.neo.openapi.utils.annotation.*;
import com.neo.v1.cas.type.*;
import javax.validation.constraints.*;
import com.fasterxml.jackson.databind.annotation.*;

import lombok.AllArgsConstructor;
import lombok.Builder;

/**
* Auto generated by ila/neo 1.2.0
*/
@Builder
@AllArgsConstructor
public class Configuration  {

	@NotNull
	private String param;

	@NotNull
	private String type;

	@NotNull
	private String value;

	@NotNull
	private String description;


  public Configuration() {
    super();
  }

  /* Getters */

	public String getParam() { return this.param; }

	public String getType() { return this.type; }

	public String getValue() { return this.value; }

	public String getDescription() { return this.description; }


  /* Setters */

	public void setParam(String param) { this.param = param; };

	public void setType(String type) { this.type = type; };

	public void setValue(String value) { this.value = value; };

	public void setDescription(String description) { this.description = description; };


  public static class Builder {

    private Configuration item = new Configuration();

		public Builder withParam(String param) { item.setParam(param); return this; };

		public Builder withType(String type) { item.setType(type); return this; };

		public Builder withValue(String value) { item.setValue(value); return this; };

		public Builder withDescription(String description) { item.setDescription(description); return this; };


    public Configuration build() {
      return item;
    }

  }


}