# cas-spec/target/generated-sources/com/neo/v1/cas/model/Meta.java
package com.neo.v1.cas.model;

import com.neo.openapi.utils.serialization.*;
import com.neo.openapi.utils.annotation.*;
import com.neo.v1.cas.type.*;
import javax.validation.constraints.*;
import com.fasterxml.jackson.databind.annotation.*;

import lombok.AllArgsConstructor;
import lombok.Builder;

/**
* Auto generated by ila/neo 1.2.0
*/
@Builder
@AllArgsConstructor
public class Meta  {

	@NotNull
	private String code;

	@NotNull
	private String message;


  public Meta() {
    super();
  }

  /* Getters */

	public String getCode() { return this.code; }

	public String getMessage() { return this.message; }


  /* Setters */

	public void setCode(String code) { this.code = code; };

	public void setMessage(String message) { this.message = message; };


  public static class Builder {

    private Meta item = new Meta();

		public Builder withCode(String code) { item.setCode(code); return this; };

		public Builder withMessage(String message) { item.setMessage(message); return this; };


    public Meta build() {
      return item;
    }

  }


}